Failover of PostgreSQL primary server
This commits extends the logic in unavailable label. We start to check if there
is an appropriate a failover target, that is sufficently caught up HA standby.
If there is such failover target, we fence the primary node by destroying it
and promote the standby.

There is a special check for async replication; we only trigger failover if the
LSN difference between primary and standby is less than 80MB (~5 WAL files).
This puts a limit on data loss for async replication. We don't need such check
for sync replication, because it should be in same LSN with primary anyway.