Start using dnsmasq as the dns server
So far we have not been using dnsmasq for DNS queries, we just used it
to configure public dns server addresses in the VM. Therefore, all the
dns queries hit to a public dns server after the VM, for sure. Now, we
configure dnsmasq to also run as a dns server. This way, if the address
is cached locally in the dnsmasq, we skip going to a public server and
respond locally from the host. To be able to do that, we assign a
private ipv6 address (fd00:b1c:100d:53:: so, ubicloud-53) to the dnsmasq
server, make it listen there and as part of the DHCP requests, we send
the dnsmasq's address to the client as the dns server.

While doing above, I realized dnsmasq listens in all the available
address space when we configure listen-address together with interfaces.
For it to specifically bind to the addresses and interfaces, we are also
adding "bind-interfaces".