Add set_lifecycle_policy method for Minio client
MinIO allows setting lifecycle policies for automatic expiration of objects in
it, which is quite useful because it saves us from implementing the logic for
tracking object expirations and purging. Lifecyles can be used with filters and
tiering, but we don't need those right now, only expiration is enough. Thus, I
did not add code for specifying filters or setting up tiering.