Set lifecycle policy for PostgresTimeline entities
We need to keep backup files for 7 days, hence we are setting up an expiration
lifecycle policy for 8 days. This only ensures the objects are deleted from the
blob storage, but it does not delete the bucket. Still it is a big improvement
from status quo because it saves lots of space.