Add run_query utility for PostgresServer
We run SQL queries on the target user database in lots of different places, all
in very similar ways. This commit introduces run_query utility to reconcile
those calls. As a side benefit it simplifies some of the tests since we don't
need to mock sshable anymore.

More subtle, but actually quite important change is that the new utility does
not use sudo to run psql as postgres user. It doesn't need to. This is thanks
to our earlier change of adding ubi user to pg_ident.conf and allow it to use
postgres user while connecting to the database. Does this weaken our security
posture? No, because ubi user was always in the sudoers list can can switch to
postgres OS user anyway.