Add log message to MinioServer health checks when exception thrown
At the moment, we blindly return false and assume the server is down for
MinioServer healthchecks when there is an exception. However, this may
not mean the server is down. Especially considering we have made some
changes to the client, enabled SSL on the Minio side recently, there is
not a clear way to understand what is going on if we simply swallow the
error. This commit adds a clog message for that scenario.