Store serial.log to the host for failed or timed out GH Runners
This commit is made for debugging purposes. When a job fails and
customers suspect the platform, we need a way to read the serial.log.
That's why, we spare a directory in the host to move the serial.log file
when a job fails. We also store serial.log and run journalctl against
the runner-script in case workflow_job is nil. This is to understand if
there was something wrong with the runner-script and it caused the job
to not be assigned or simply the job is canceled.

We also change the nap time from 0 to 15 for exited runner-scripts. The
main reason is that, we detect the runner is exited but we don't know
if it's succeeded or failed. This information is only populated when we
get the webhook from github. Therefore, we give github 15 seconds to
update us, otherwise, probably the conclusion field will be "InProgress"
which we will filter out and not store the journal and serial.log