Increase runner recycle threshold from 2 to 5 minutes
Our runners don't know which task to execute until it's assigned. If a
workflow task is cancelled before a runner is assigned, we're unsure
which runner to remove. Therefore, if no task is assigned to a runner
within two minutes, we presume there's no task pending and proceed to
destroy the runner. Recently, we've noticed that GitHub isn't able to
assign tasks to runners within two minutes sometimes. As a result, we've
decided to extend this threshold to five minutes. The downside is that
these idle runners occupy our core capacity for an additional three
minutes. However, this situation is relatively rare, and we don't charge
customers for idle runners.