Logout of all sessions after changing email or password
It's advisable to sign out of all active sessions for the current
account after changing the email or password.

Let's say the user's password has been somehow compromised and an
attacker has accessed the user account. Even if the user resets the
password, the attacker's session remains active on their system.