Use randomized file name for MinIO CA
We place MinIO root certificates under var/ca_bundles. We used to use the
access_key as name for CA bundle, but access_key can be reused accross
clusters. With this commit, we start to hash the file content and use that as
the name. It both avoids conflicts and ensures that files would be rewritten
when the content changes (i.e. when we rotate the root certs).