Add GithubRunner to monitor
Previously we only monitored PostgresServer and MinioServer entities. It would
be beneficial to monitor GithubRunner entities as well, so that we can get an
early alert if something goes unexpected. As a small bonus, we are collecting
free memory amount as well, although we don't record it anywhere. It will be
only written to logs.

Currently, we do not try to self-heal or page. Primary purpose is logging and
manual debugging.