Enable billing for PostgreSQL standbys
There are few alternative ways for billing PostgreSQL standbys, but none are
perfect.

One option was multiplying the amount with number of standbys, but then, there
is no way to distinguish between 8 vCPU backed database from 4 vCPU backed and
HA enabled database.

Another option was adding a count to the billing records, but that would need
a bigger refactoring that I'm not too keen on right now. Though, I noted this
requirement for billing system v2.

Simple and good enough method is adding new billing rates for standbys. This
also has few extra benefits such as it allows us to charge different rate for
standbys if we want.