Add an option to enable swap for a vm
We aim to enable swap space for certain virtual machines, such as GitHub
runner ones. cloudinit offers two methods for setting up swap[^1]: the
'mounts' configuration and the 'swap' configuration. I've chosen to use
the 'swap' configuration since it's more straightforward.

I save the desired swap size to the Nexus strand's stack and pass it to
the virtual machine through 'prep.json'. If no value is provided, we
won't setup a swap.

[^1]: https://cloudinit.readthedocs.io/en/latest/reference/examples.html#adjust-mount-points-mounted