Enable swap space for GitHub runners
Recently, we've noticed some OOM events occurring with our GitHub
runners. Sometimes, these events even result in the termination of the
ssh-agent or the runner-script. After reviewing the GitHub-hosted
runners, we found that they have 4294963200 bytes of swap space enabled.
We've decided to implement the same amount of swap space in our system.

Fixes #1217