Specify node and npm versions in package.json
This is what the Heroku build pack reads to determine what version to
use, as-is we have spurious difference between development and
production.

Clover only uses node for compiling css, so, it was not of major
concern, but reading build pack output closely showed Heroku was
warning us about a few things that it seemed like a good idea to get
around to.