For PG health checks catch exceptions thrown while opening connection
When Sequel.connect is out of the block, raised exceptions leak to the monitor
level and evaluated as permanent failure. With this change we start to count
those as down events instead of permanent failures. This is achieved by moving
the Sequel.open statement inside the block that already has rescue section.