Add CloverError as base error for errors on routing
Before the commit, all different errors were created by extending
the StandardError and parse_error was responsible to create error
responses dependent on the passed error object. Though, that means
with the addition of new errors in the future parse_error can be
unnecessarily overloaded.

So, adding a base class and extending that for the errors faced while
routing the request. So, the parse_error will only discriminate
between clover errors and other errors from roda or sequel. Class
that extends the base class will keep the common information preset
and expects remaining ones from the code creating the error.