Add concession table to keep resource type level discount and credit
In order to keep concession on the resource type level (sku is also used
interchangeably at some points) adding a new table concession.

Multiple concessions can be added for the same resource type. If the
concession tuple only has project id column but not resource type,
that means the concession will be applicable for all resources
within that project.