Update invoice generator to apply resource type level concession
Updating invoice generator code to apply concessions added to
the concession table. Existing invoices and credit/discount
info on the project table are still supported. In other words,
concessions added to concession table will be applied together
with the credit/discount values added to project table. Structure
of the invoice will be updated with a subsequent PR and credit/discount
info on the project table will also be removed with another one.

Concessions will be applied in the order of

1- Apply resource type level concessions
	- Apply discounts (by multiplying if multiple defined for same type)
	- Apply credit

2- Apply project level concessions
	- Apply discounts defined on concession table
	- Apply discount defined on project table
	- Apply credits defined on concession table
	- Apply credit defined on project table

Also note that, to not affect the invoice structure with this PR,
resource type level values are added first for simpler calculations
then removed before creating the invoice object.