Start using nftables at host to block unused ip addresses
We recently got an e-mail from Hetzner saying our hosts produce a
routing loop with their gateway when receiving packets for IPs not yet
allocated to a VM.

In this case, the host will forward the packet to the default route,
which is the gateway, which in turn, forwards it back to the host,
until TTL is decremented, with much resource expenditure by both
parties.

To address this, I introduce a new nftables definition at the host
level, with the intention of dropping packets to IP addresses with no
VM allocating them. I add a new table of nftables that has two sets
and two rules:
1. The first rule is to accept packets if the daddr is part of the
   first set.
2. The second rule is to drop packets if the daddr is part of the
   second set.

This commit simply sets up the hosts with necessary path creations and
some changes in the /etc/nftables.conf file to accommodate that.

A small note regarding IPv6: We don't need to implement a similar
mechanism for IPv6, because the behavior is different there. When an
IPv4 packet reaches us and we don't know what to do, we simply send it
back to the network from the default route. In IPv6 on the other hand,
we send a neighbor solicitation message to the local network, if there
is a match, they would respond with a link local address, if there is
not, we don't get that response and send back to the client a
destination unreachable message.