Start making nftables changes for VM de/provisionings
Now, we start updating the nftables definitions at the VM creation and
deletion times. Though, these are still not activated since we did not
enable nftables.service. We want to make sure the VMs are creating the
necessary files and removing at the time of deprovisioning properly.
Also, this will give us time to graciously wait for the existing VMs
without introducing a connectivity problem. Once all of the VMs are
rotated and the files are created manually for the long running VMs, we
may proceed to the next commit.