Use channel_binding=require in connection string
Although there is a CA apparatus set up for databases, it's not
exposed in the UI anywhere.  Thus, unless one is unusually
Postgres-aware, one is likely to use the database in a manner that
will be prone to MITM attack.  Also, even with a trust root available,
people tend to connect in a way that is prone to MITM attacks anyway:
it's a bit of a hassle to set up.

As it turns out, since Postgres 13's libpq, one can set the
`channel_binding` to `require` which will ensure both the server and
client have the same view of the communication channel's cryptographic
information before connecting, preventing MITM attacks.  Postgres 13
was long enough ago that I think we can inconvenience people with even
older libpq.