Rescue access errors while fetching backups
Sometimes the access configuration is not ready yet when we try to fetch
backups from blob storage. In those cases, it is better experience to just
return an empty backup list instead of an exception. We already catch one such
exception but it seems depending on the process in access configuration, MinIO
generates different error messages. With this commit we start to catch another
such exception.