Add extra data and severity to pages
The `severity` field allows us to designate the alert's severity level.
PagerDuty uses this field to assess the alert's urgency. The severity
field can be assigned one of the following values: `critical`, `error`,
`warning`, `info` [^1]. In the past, we assigned the `error` value to
all alerts.

Additionally, I added the `extra_data` parameter. This facilitates the
addition of extra details to incidents, along with related resources.
For instance, we can include utilization in capacity incidents.

While I usually appreciate the splat operator `*tag_parts`, its
comprehension becomes challenging as the number of parameters increases.
Therefore, I modified it to accept a list of tags.

[^1]: https://support.pagerduty.com/docs/dynamic-notifications#severity-and-urgency-mapping