Fix MINIO_SERVER_URL config for clusters without DnsZone
When we refactored minio.urls at
https://github.com/ubicloud/ubicloud/pull/1181/files we overlooked the
development minio cluster provisioning without dns_zone.
MINIO_SERVER_URL configuration is not necessary if we do not have a dns
name or a load balancer in front of the cluster. Furthermore, minio
requires this configuration to be the same in all of the servers in the
cluster, if not, it rejects to spin up. This causes a problem for
development environment because we do not have dnszone setup in dev.
This PR simply implements a quick fix for that.