Fix the reusing pooled vm when the runner doesn't wait the vm
The runner started to not waiting its vm destroy at
ee8ec26247274964853eacd9435253fa36cbfedc. After deployment, another
issue surfaced. I discovered that one of the runners was at
`register_runner` label, but its VM had been deleted already. Furkan
pointed that the VM pool selects a VM that's running state and doesn't
have a GithubRunner model. Therefore, if the runner model doesn't wait
for the VM to be destroyed, the pool may select a VM that's already been
used because its GithubRunner has been deleted.

This PR simply updates the display state of the vm as `deleting` while
destroying the runner, so it will not picked again.