Update runner user with runneradmin and add new runner user
Renaming the unix user name of github runner vms from runner to runneradmin
then adding a new user with the name runner. Since the runneradmin will be
the first user created on vm, it's uid will be 1000 and as the runner
user will be the next one, it's uid will be 1001.

Having both runneradmin and runner users with uid 1000 and 1001 will make
the users on the Ubicloud same with what Github has on standard runners.

Please note, those uids won't be same with the Github's larger runners
as Github has discrepancy between standard and larger runners. Check the
issue https://github.com/actions/runner-images/issues/6930 for details.