Add endpoint to restart PostgreSQL database
While being able to restart your own database is useful on its own, it is also
a requirement for being able to modify(and see its effect) of some GUCs. For
example, currently, we are adding new extensions and some of those extensions
require to be loaded via shared_preload_libraries GUC, which requires restart.