Limit monitor threads
We create 2 threads per resource. If the number of resources is high, it is
possible to hit max thread limits. In medium term, we should either start to
use threads more efficiently (by using something like fiber) or implement
support for multiple monitors.

Until then, we are adding this limit, so that the monitor process wouldn't
crash due to lack of resources. This means we are not monitoring all resources,
but it is still better than having a crash and not being able to monitor none
of the resources. In this way,
- If the resource already has a thread, it continues to use that thread. We
don't stop to monitor the resource.
- If a resource cannot get a thread, we retry in 5 minutes and hopefully it can
get a thread then.

This method prioritizes monitoring long-running resources, which are usually
more critical anyway. Still, it is not ideal and we should find a way to
monitor all resources (such as using fibers for sub-thread processing).