Get optional tax_id and company name on billing page
Adding support to get tax_id and company name information from
customers on the billing page, so we can show those on invoices.

Note that, we are getting tax id and company name from customer after
they add billing info to their project via Stripe. We can not get those
informations from customers while creating Stripe Checkout Session as
Stripe does not allow getting them with the way we are using it. We are
keeping them as metadata on the Stripe, so we don't have them on our
database.