Send E2E failed notification for timed out runs too
GitHub displays `canceled` instead of `failed` for workflows that time
out.  We don't receive a notification for timed out workflow since
`failure()` is false for them. `!success()` returns true for canceled
runs too. However, it has a drawback: if we manually cancel a run on the
main branch, it will still send a notification.

We decided to failed our job if it times out. We start the foreman
script with simple `timeout` command. If the script doesn't finish in
time, the `timeout` command will kill it and return non-zero exit code.
We catch this exit code and fail the job.