Fix destroy of strand who has multiple frames in stack
When a strand has multiple frames in its stack, and it is destroyed, the top
frame destroys the associated entities and pops. Then next frame continues to
work, but fails because the entities are already destroyed. This is fixed by
first popping all frames but the last one, and then switching to the destroy
label to handle cleanup.