Add migration for GithubRepository model
GitHub provides an API endpoint to access installation repositories.
However, not all customers use Ubicloud runners for all their
repositories. While we have information of active repositories in the
DeletedRecord, we prefer not to use it for application logic. Therefore,
I've introduced the GithubRunner model to maintain a list of active
repositories. This model is created when the first job is delivered for
the repository. It also includes a 'last_job_at' column to help to clean
up inactive repositories. Our job polling mechanism need active
repositories.