Block tcp udp port 111
We recently got an e-mail from a cyber security agency stating that one
of the VMs have port 111 reachable from the public internet. Looks like
this port is used by Portmapper (or RPCbind) which are used for various
tasks like network attached file system. We are not making use of these
services and there is possibly not a good reason why a customer would
need. When left reachable, this port poses a security risk that can be
exploited for DDoS reflection attacks. Therefore, all things considered,
we have decided to block port 111 by default.