Box virtual networking encryption key in stdin
Generally, we keep secrets that need to be passed to commands out of
the sudo audit logs / the systemd journal and clover log
telemetry/Clog output by boxing them inside stdin.  We missed one
opportunity to do that here.

Good news, though: NIC rekeying is so frequent that, a day after we
deploy this, all old secrets will be obsolete.  Reference this code in
`vnet/subnet_nexus.rb`:

    if private_subnet.last_rekey_at < Time.now - 60 * 60 * 24
      private_subnet.incr_refresh_keys
    end