Add application type checking to heartbeat
A trick for more robust ditch-effort monitoring is to ensure at least
one backend for each expected process type is connected to Postgres.

In principle this could be made stricter by checking for the quantity
of each backend relative to expectations, but these expectations are
quite unstable given changes in pooling and compute acquisition, so
it's more suited to a not-yet-designed Prog that would offer something
more rich than what "heartbeat" is intended to provide.