Update rubocop and ignore new rails cop
I originally tried to use this cop and let it rewrite the source, but
it rewrites it to use symbols that are not defined unless one has
Rails loaded, and Clover is not a Rails application.

So, ignore the new cop and leave things the way they are.

However, we do want to update rubocop so our dependabot updates can
clear CI as they usually do, and CI depends on rubocop exiting with no
corrections offered.