Add GithubRepository nexus to poll queued jobs
Our integration with GitHub Actions runner relies on the webhook
delivered by GitHub. If GitHub fails to deliver the
"workflow_job.queued" event, our control plane remains unaware of the
new queued job and cannot provision a runner for it. If GitHub attempts
and fails to deliver, we have "Prog::RedeliverGithubFailures" to retry
the delivery. However, if GitHub doesn't attempt delivery, we have no
way of knowing. We've encountered instances where GitHub failed to
deliver the "workflow_job.queued" event, and we currently have an active
support ticket for this issue. GitHub engineers are working to resolve
this. Furthermore, GitHub has had 6 incidents in the past 7 days. We
need to develop a solution to make our service more resilient to
GitHub's issues.

We're also addressing an issue related to the "max-parallel"
configuration. GitHub Actions workflow syntax allows to limit concurrent
jobs in a matrix via the "max-parallel" configuration. Unfortunately, it
sends the "workflow_job.queued" event for all jobs in the matrix
simultaneously. The job payload lacks an indicator for whether the job
is limited by "max-parallel" or not. We don't have this indicator in the
API payload either. The poller views all jobs as queued. It provisions a
runner for all jobs in the matrix until the workflow run is finished
when the initially provisioned runner is recycled. However, the runner
will pick jobs from the queue one by one based on the "max-parallel"
value. There are several issues about "max-parallel" config design, but
GitHub doesn't want to improve it. I think "max-parallel" is misdesigned
and has architectural problem, but yeah we need to deal with it. The
poller is our best solution to this problem, even if it means keeping
idle runners for limited jobs.

The poller design presents its own challenges. Let's first understand
GitHub's resources. We have a GitHub app that customers install on their
accounts. We keep a GithubInstallation record for each customer.
Customers may allow only certain repositories or all repositories while
installing our GitHub app. Each repository has workflow runs, and each
run has workflow jobs. When we receive a "workflow_job.queued" event, we
create a GithubRunner record if the payload contains the "ubicloud"
label.

There isn't an endpoint that provides queued jobs for a GitHub
installation with the "ubicloud" label. We need to retrieve all
repositories for each GitHub installation, get queued workflow runs for
each repository, get jobs for each run, and check the job's status and
label. This process requires numerous requests. We've decided to poll
only active repositories. If no new job is queued for a repository in
the last 6 hours, we don't poll it. In other words, at least one job
should arrive within a 6-hour timeframe for us to poll the repository.

Furthermore, self-hosted runners are job-agnostic. We can't register a
runner for a specific job. We register with the requested label to the
repository, and GitHub assigns the job to the appropriate runner. Thus,
the poller compares the count of queued jobs and idle runners for the
repository.

"Prog::Github::RepositoryNexus" polls the repository's jobs every 5
minutes. The "max-parallel" scenario requires more frequent polling than
the "missing webhook" scenario. I chose 5 minutes interval for the
poller because our runner recycling threshold is 5 minutes.
So for "max-parallel" scenarios, we will keep idle runners for 5
minutes. No need to provision additional runners for 5 minutes. For
missing webhook events, 5 minutes is a reasonable interval.