Update StorageDevice.available_storage_gib at Vm.destroy
Looks like we completely forgot to update available_storage_gib of the
storage_device at the Vm.destroy. While this is not harmless at the
moment, since we do not use this property anywhere, it will be used soon
with the changes in https://github.com/ubicloud/ubicloud/pull/1281.
Therefore, we must fix this before deploying those changes.