Pick host for VM randomly
We used to pick the host for a VM based on memory/CPU ratio. This causes
predominantly picking the same host(s) if there are some hosts with different
memory/CPU ratios than rest of the fleet.

This is especially easy to analyze at present because we don't have multiple
memory tiers available to the customer just yet, otherwise it'd do something
weird like put a memory heavy provision on something with little memory,
leaving very little room for anything else. Though we have host memory ratio
skew, it nevertheless does not translate into the customer's bought capacity
for anything we sell at present, so we are eating the mismatch for now.

For example a recent pathological case was that we had new hosts with 1GBit
network cards, and the rest of the fleet had 10GBit network cards. The new
hosts were picked for VMs, because they had different memory/CPU ratios.

We are also working on more sophisticated host picking algorithm, but that item
got delayed for the sake of higher priority tasks. I think this change would
improve the situation in the meantime.