Add prog to update ip addresses of globally blocked dns names
This commit introduces a new prog that will periodically check the ip
addresses of globally blocked dnsnames. This way, the list will be kept
fresh and the new VMs will be provisioned with the new list. In future,
we might introduce a system here to trigger a firewall rule update for
existing VMs as well.