Fix nil ip_list issue in globally_blocked_dnsname
The commit 540b34854f6ba13b1e546b5741b4523c50b7bd62 introduced an
iteration via .each over the ip_list column of globally_blocked_dnsname.
However, when we add a new entry to the table, until the first iteration
of the Prog::ResolveGloballyBlockedDnsnames, the ip_list is empty.
Therefore, the Prog::UpdateFirewallRules starts to fail with nil ref.
This commit fixes that.