Add a program to remove an spdk installation
Current spdk upgrade path is to install a new spdk version and then
remove the older version. We could install a new version using
`Prog::Storage::SetupSpdk.assemble(host_id, version)`, but a program to
remove an older version was missing.

This PR adds that. This program disables the installation by setting its
allocation_wait to 0, then waits until volumes using that installation
has been destroyed, and then removes it.

This is supposed to be invoked manually by an operator. To invoke it,
one should do:

```
> Prog::Storage::RemoveSpdk.assemble(spdk_installation_id)
```